fix(subgrid): cast fine-source scatter to dest dtype, closes float64->float32 x64 leak (#485)
_inject_fine_sources scatters the per-step waveform value into the fine
E-field arrays with .at[...].add(...). init_state pins those arrays at
float32 (its field_dtype default is x64-independent), but the waveform
array is built via jnp.array(...) which DOES track x64 -- float64 once
jax.experimental.enable_x64() is scoped on. Under scoped x64 that mixed
dtypes at the scatter site, which JAX flags with a FutureWarning ("cannot
safely cast value from dtype=float64 to dtype=float32 ... In future JAX
releases this will result in an error") -- non-fatal today, fatal on a
future JAX upgrade. Same class as #14 and #483.

Fix: cast the source value to the destination array's own dtype
(.astype(ez_arr.dtype) / ex_arr.dtype / ey_arr.dtype) explicitly before
the scatter-add. No-op under the float32 default (verified bit-identical
time_series against pre-fix output on the excited-port subgrid fixture);
an explicit, safe downcast under x64, so JAX no longer treats it as an
unsafe implicit promotion.

New test test_subgrid_source_injection_dtype.py verified both directions
locally: FAILS on pre-fix code (FutureWarning raised at
jit_runner.py:1758 inside _inject_fine_sources) and PASSES on the fix.
Full subgrid suite: 114 passed before -> 116 passed after (2 new), same
4 deselected / 2 xfailed / warnings -- no regressions.

_inject_coarse_sources has the identical unfixed pattern (src_vals_c);
out of scope here since #485 named only the fine-source path.

R3: memory=none-after-grep (rfx-known-issues.md, index.md; precedent is
simulation.py:816-822's jnp.result_type dtype-tracking fix for #477/#483,
consistent not contradicted) | R2-attempts=0 (single clean attempt,
falsifier pre-declared and confirmed both directions) | falsifier=rerun
test_subgrid_source_injection_dtype.py::test_fine_source_injection_no_dtype_futurewarning_under_x64
with warnings.simplefilter("error", FutureWarning) under scoped x64

+++ rfx/subgridding/jit_runner.py
@@ -1753,13 +1753,21 @@ def _make_step_fn(ctx):
         ex_f, ey_f, ez_f, hx_f, hy_f, hz_f = carry["f"]
 
         def _inject_fine_sources(ex_arr, ey_arr, ez_arr):
+            # src_vals is built from jnp.array(waveform) (jit_runner.py ~2615),
+            # which tracks x64: float32 with x64 off (matches ex/ey/ez_arr's
+            # init_state default), float64 once x64 is scoped on. ex/ey/ez_arr
+            # stay float32 regardless (init_state's field_dtype default), so
+            # the x64-on scatter mismatches dtypes and JAX raises a
+            # cast-safety FutureWarning (issue #485). Cast explicitly to the
+            # destination array's own dtype: a no-op under the float32
+            # default (bit-identical), and a safe explicit downcast under x64.
             for idx_s, (si, sj, sk, sc) in enumerate(src_meta):
                 if sc == "ez":
-                    ez_arr = ez_arr.at[si, sj, sk].add(src_vals[idx_s])
+                    ez_arr = ez_arr.at[si, sj, sk].add(src_vals[idx_s].astype(ez_arr.dtype))
                 elif sc == "ex":
-                    ex_arr = ex_arr.at[si, sj, sk].add(src_vals[idx_s])
+                    ex_arr = ex_arr.at[si, sj, sk].add(src_vals[idx_s].astype(ex_arr.dtype))
                 elif sc == "ey":
-                    ey_arr = ey_arr.at[si, sj, sk].add(src_vals[idx_s])
+                    ey_arr = ey_arr.at[si, sj, sk].add(src_vals[idx_s].astype(ey_arr.dtype))
             return ex_arr, ey_arr, ez_arr
 
         def _inject_coarse_sources(ex_arr, ey_arr, ez_arr):
